Split mmap.rs into one-pub-per-file module structure
Refactors src/mmap.rs into src/mmap/ directory:

- mapped_archive.rs: MappedArchive struct (read-only)
- mapped_archive_mut.rs: MappedArchiveMut struct (read-write)
- mod.rs: re-exports

Tests are co-located with their respective structs.
